@@ -1,27 +1,12 @@
-import React from 'react';
-import logo from './logo.svg';
-import './App.css';
-import * as wasm from "wasm-solver";
-
+import React from "react";
+import Header from "./components/header/Header";
+import Main from "./pages/main/Main";
 
 function App() {
-  wasm.greet();
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.tsx</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+    <div className="flex flex-col h-full">
+      <Header />
+      <Main />
     </div>
   );
 }
